+++ client/guilder-client-template/src/client.rs
@@ -1,4 +1,4 @@
-use guilder_abstraction::{self, Result<bool, String>, Result<i64, String>, Result<Vec<String>, String>, Result<f64, String>, L2Update, Fill, AssetContext, Liquidation};
+use guilder_abstraction::{self, Result<bool, String>, Result<i64, String>, Result<Vec<String>, String>, Result<Decimal, String>, L2Update, Fill, AssetContext, Liquidation};
 use futures_util::stream;
 use reqwest::Client;
 
@@ -32,11 +32,11 @@ impl guilder_abstraction::GetMarketData for ExchangeClient {
         unimplemented!()
     }
 
-    async fn get_price(&self, symbol: String) -> Result<f64, String> {
+    async fn get_price(&self, symbol: String) -> Result<Decimal, String> {
         unimplemented!()
     }
 
-    async fn get_open_interest(&self, symbol: String) -> Result<f64, String> {
+    async fn get_open_interest(&self, symbol: String) -> Result<Decimal, String> {
         unimplemented!()
     }
 
@@ -45,11 +45,11 @@ impl guilder_abstraction::GetMarketData for ExchangeClient {
 #[allow(unused_variables)]
 #[allow(async_fn_in_trait)]
 impl guilder_abstraction::ManageOrder for ExchangeClient {
-    async fn place_order(&self, symbol: String, price: f64, volume: f64) -> Result<i64, String> {
+    async fn place_order(&self, symbol: String, price: Decimal, volume: Decimal) -> Result<i64, String> {
         unimplemented!()
     }
 
-    async fn change_order_by_cloid(&self, cloid: i64, price: f64, volume: f64) -> Result<i64, String> {
+    async fn change_order_by_cloid(&self, cloid: i64, price: Decimal, volume: Decimal) -> Result<i64, String> {
         unimplemented!()
     }
